+++ server/src/schema/Types/KarbanProjectTabCardType.ts
@@ -1,0 +1,9 @@
+import { GraphQLObjectType, GraphQLString } from 'graphql';
+
+export const KarbanProjectTabCardType = new GraphQLObjectType({
+  name: 'KarbanProjectTabCard',
+  fields: () => ({
+    cardId: { type: GraphQLString },
+    cardBody: { type: GraphQLString },
+  }),
+});
